Cleans up deepmass
