+++ jax_lensing/models/__init__.py
@@ -1,0 +1,2 @@
+from jax_lensing.models.convdae import UResNet18
+from jax_lensing.models.normalization import SNParamsTree, SpectralNorm
