--- jax_lensing/models/convdae2.py
@@ -1,299 +1,0 @@
-import jax
-import jax.numpy as jnp
-import haiku as hk
-
-import types
-from typing import Mapping, Optional, Sequence, Union
-
-def check_length(length, value, name):
-  if len(value) != length:
-    raise ValueError(f"`{name}` must be of length 4 not {len(value)}")
-
-class BlockV1(hk.Module):
-  """ResNet V1 block with optional bottleneck."""
-
-  def __init__(
-      self,
-      channels: int,
-      stride: Union[int, Sequence[int]],
-      use_projection: bool,
-      bn_config: Mapping[str, float],
-      bottleneck: bool,
-      transpose: bool = False,
-      name: Optional[str] = None
-  ):
-    super().__init__(name=name)
-    self.use_projection = use_projection
-
-    bn_config = dict(bn_config)
-    bn_config.setdefault("create_scale", True)
-    bn_config.setdefault("create_offset", True)
-    bn_config.setdefault("decay_rate", 0.999)
-
-    if transpose:
-      maybe_transposed_conv = hk.Conv2DTranspose
-    else:
-      maybe_transposed_conv = hk.Conv2D
-
-    if self.use_projection:
-      self.proj_conv = maybe_transposed_conv(
-          output_channels=channels,
-          kernel_shape=1,
-          stride=stride,
-          with_bias=False,
-          padding="SAME",
-          name="shortcut_conv")
-
-      self.proj_batchnorm = hk.BatchNorm(name="shortcut_batchnorm", **bn_config)
-
-    channel_div = 4 if bottleneck else 1
-    conv_0 = hk.Conv2D(
-        output_channels=channels // channel_div,
-        kernel_shape=1 if bottleneck else 3,
-        stride=1,
-        with_bias=False,
-        padding="SAME",
-        name="conv_0")
-    bn_0 = hk.BatchNorm(name="batchnorm_0", **bn_config)
-
-    conv_1 = maybe_transposed_conv(
-        output_channels=channels // channel_div,
-        kernel_shape=3,
-        stride=stride,
-        with_bias=False,
-        padding="SAME",
-        name="conv_1")
-
-    bn_1 = hk.BatchNorm(name="batchnorm_1", **bn_config)
-    layers = ((conv_0, bn_0), (conv_1, bn_1))
-
-    if bottleneck:
-      conv_2 = hk.Conv2D(
-          output_channels=channels,
-          kernel_shape=1,
-          stride=1,
-          with_bias=False,
-          padding="SAME",
-          name="conv_2")
-
-      bn_2 = hk.BatchNorm(name="batchnorm_2", scale_init=jnp.zeros, **bn_config)
-      layers = layers + ((conv_2, bn_2),)
-
-    self.layers = layers
-
-  def __call__(self, inputs, is_training, test_local_stats):
-    out = shortcut = inputs
-
-    if self.use_projection:
-      shortcut = self.proj_conv(shortcut)
-      shortcut = self.proj_batchnorm(shortcut, is_training, test_local_stats)
-
-    for i, (conv_i, bn_i) in enumerate(self.layers):
-      out = conv_i(out)
-      out = bn_i(out, is_training, test_local_stats)
-      if i < len(self.layers) - 1:  # Don't apply relu on last layer
-        out = jax.nn.relu(out)
-
-    return jax.nn.relu(out + shortcut)
-
-class BlockGroup(hk.Module):
-  """Higher level block for ResNet implementation."""
-
-  def __init__(
-      self,
-      channels: int,
-      num_blocks: int,
-      stride: Union[int, Sequence[int]],
-      bn_config: Mapping[str, float],
-      bottleneck: bool,
-      use_projection: bool,
-      transpose: bool,
-      name: Optional[str] = None,
-  ):
-    super().__init__(name=name)
-
-    block_cls = BlockV1
-
-    self.blocks = []
-    for i in range(num_blocks):
-      self.blocks.append(
-          block_cls(channels=channels,
-                    stride=(1 if i else stride),
-                    use_projection=(i == 0 and use_projection),
-                    bottleneck=bottleneck,
-                    bn_config=bn_config,
-                    transpose=transpose,
-                    name="block_%d" % (i)))
-
-  def __call__(self, inputs, is_training, test_local_stats):
-    out = inputs
-    for block in self.blocks:
-      out = block(out, is_training, test_local_stats)
-    return out
-
-
-class UResNet(hk.Module):
-  """ Implementation of a denoising auto-encoder based on a resnet architecture
-  """
-
-  def __init__(self,
-               blocks_per_group,
-               bn_config,
-               bottleneck,
-               channels_per_group,
-               use_projection,
-               name=None):
-    """Constructs a Residual UNet model based on a traditional ResNet.
-    Args:
-      blocks_per_group: A sequence of length 4 that indicates the number of
-        blocks created in each group.
-      bn_config: A dictionary of two elements, ``decay_rate`` and ``eps`` to be
-        passed on to the :class:`~haiku.BatchNorm` layers. By default the
-        ``decay_rate`` is ``0.9`` and ``eps`` is ``1e-5``.
-      resnet_v2: Whether to use the v1 or v2 ResNet implementation. Defaults to
-        ``False``.
-      bottleneck: Whether the block should bottleneck or not. Defaults to
-        ``True``.
-      channels_per_group: A sequence of length 4 that indicates the number
-        of channels used for each block in each group.
-      use_projection: A sequence of length 4 that indicates whether each
-        residual block should use projection.
-      name: Name of the module.
-    """
-    super().__init__(name=name)
-    self.resnet_v2 = False
-
-    bn_config = dict(bn_config or {})
-    bn_config.setdefault("decay_rate", 0.9)
-    bn_config.setdefault("eps", 1e-5)
-    bn_config.setdefault("create_scale", True)
-    bn_config.setdefault("create_offset", True)
-
-    # Number of blocks in each group for ResNet.
-    check_length(4, blocks_per_group, "blocks_per_group")
-    check_length(4, channels_per_group, "channels_per_group")
-
-    self.initial_conv = hk.Conv2D(
-        output_channels=32,
-        kernel_shape=7,
-        stride=2,
-        with_bias=False,
-        padding="SAME",
-        name="initial_conv")
-
-    if not self.resnet_v2:
-      self.initial_batchnorm = hk.BatchNorm(name="initial_batchnorm",
-                                            **bn_config)
-
-    self.block_groups = []
-    self.up_block_groups = []
-    strides = (1, 2, 2, 1)
-    for i in range(4):
-      self.block_groups.append(
-          BlockGroup(channels=channels_per_group[i],
-                     num_blocks=blocks_per_group[i],
-                     stride=strides[i],
-                     bn_config=bn_config,
-                     bottleneck=bottleneck,
-                     use_projection=use_projection[i],
-                     transpose=False,
-                     name="block_group_%d" % (i)))
-
-    for i in range(4):
-      self.up_block_groups.append(
-          BlockGroup(channels=channels_per_group[i],
-                     num_blocks=blocks_per_group[i],
-                     stride=strides[i],
-                     bn_config=bn_config,
-                     bottleneck=bottleneck,
-                     use_projection=use_projection[i],
-                     transpose=True,
-                     name="up_block_group_%d" % (i)))
-
-    if self.resnet_v2:
-      self.final_batchnorm = hk.BatchNorm(name="final_batchnorm", **bn_config)
-
-    self.final_upconv = hk.Conv2DTranspose(output_channels=1,
-                                kernel_shape=5,
-                                stride=2,
-                                padding="SAME",
-                                name="final_upconv")
-
-    self.final_conv = hk.Conv2DTranspose(output_channels=1,
-                                kernel_shape=5,
-                                stride=2,
-                                padding="SAME",
-                                name="final_conv")
-
-  def __call__(self, inputs, condition, is_training, test_local_stats=False):
-    out = inputs
-    out = jnp.concatenate([out, condition*jnp.ones_like(out)[...,[0]]], axis=-1)
-    out = self.initial_conv(out)
-
-    # Decreasing resolution
-    levels = []
-    for block_group in self.block_groups:
-      levels.append(out)
-      out = block_group(out, is_training, test_local_stats)
-
-    out = jnp.concatenate([out, condition*jnp.ones_like(out)],axis=-1)
-
-    # Increasing resolution
-    for i, block_group in enumerate(self.up_block_groups[::-1]):
-      out = block_group(out, is_training, test_local_stats)
-      out = jnp.concatenate([out, levels[-i-1]],axis=-1)
-
-    # Second to last upsampling, merging with input branch
-    return self.final_conv(out)/(jnp.abs(condition)*jnp.ones_like(inputs)+1e-3)
-
-class SmallUResNet(UResNet):
-  """ResNet18."""
-
-  def __init__(self,
-               bn_config: Optional[Mapping[str, float]] = None,
-               name: Optional[str] = None):
-    """Constructs a ResNet model.
-    Args:
-      bn_config: A dictionary of two elements, ``decay_rate`` and ``eps`` to be
-        passed on to the :class:`~haiku.BatchNorm` layers.
-      resnet_v2: Whether to use the v1 or v2 ResNet implementation. Defaults
-        to ``False``.
-      name: Name of the module.
-    """
-    super().__init__(blocks_per_group=(2, 2, 2, 2),
-                     bn_config=bn_config,
-                     bottleneck=False,
-                     channels_per_group=(32, 64, 128, 128),
-                     use_projection=(True, True, True, True),
-                     name=name)
-
-
-class MediumUResNet(UResNet):
-  """ResNet18."""
-
-  def __init__(self,
-               bn_config: Optional[Mapping[str, float]] = None,
-               name: Optional[str] = None):
-    """Constructs a ResNet model.
-    Args:
-      bn_config: A dictionary of two elements, ``decay_rate`` and ``eps`` to be
-        passed on to the :class:`~haiku.BatchNorm` layers.
-      resnet_v2: Whether to use the v1 or v2 ResNet implementation. Defaults
-        to ``False``.
-      name: Name of the module.
-    """
-    super().__init__(blocks_per_group=(2, 2, 2, 2),
-                     bn_config=bn_config,
-                     bottleneck=False,
-                     channels_per_group=(32, 64, 128, 128),
-                     use_projection=(True, True, True, True),
-                     name=name)
-
-def forward(x, s, is_training=False):
-    denoiser = MediumUResNet()
-    return denoiser(x, s, is_training=is_training)
-
-model = hk.transform_with_state(forward)
-
-#sn_fn = hk.transform_with_state(lambda x: hk.SNParamsTree(ignore_regex='[^?!.]*b$')(x))
-#sn_fn = hk.transform_with_state(lambda x: CustomSNParamsTree(ignore_regex='[^?!.]*b$',val=2.)(x))
