+++ scripts/train_deepmass.py
@@ -1,16 +1,14 @@
 # Script for training a DeepMass inference model
 import os
-
 os.environ['XLA_FLAGS']='--xla_gpu_cuda_data_dir=/gpfslocalsys/cuda/10.1.2'
 
 from absl import app
 from absl import flags
 
 import haiku as hk
+import optax
 import jax
 import jax.numpy as jnp
-from jax.experimental import optix
-
 import numpy as onp
 import pickle
 from functools import partial
@@ -19,13 +17,11 @@ from astropy.io import fits
 from flax.metrics import tensorboard
 
 # Import tensorflow for dataset creation and manipulation
-import tensorflow.compat.v2 as tf
-tf.enable_v2_behavior()
+import tensorflow as tf
 import tensorflow_datasets as tfds
 
-from jax_lensing.models.convdae import SmallUResNet
-from jax_lensing.models.convdae2 import MediumUResNet
-from jax_lensing.models.normalization import SNParamsTree as CustomSNParamsTree
+from jax_lensing.models.convdae import UResNet18
+from jax_lensing.models.normalization import SNParamsTree
 from jax_lensing.spectral import make_power_map
 from jax_lensing.utils import load_dataset_deepmass
 from jax_lensing.inversion import ks93, ks93inv
@@ -39,23 +35,15 @@ flags.DEFINE_string("train_split", "90%", "How much of the training set to use."
 flags.DEFINE_string("mask", "../data/COSMOS/cosmos_full_mask_0.29arcmin360copy.fits", "Path to input mask.")
 flags.DEFINE_string("std1", "../data/COSMOS/std1.npy", "Standard deviation noise e1 (gal).")
 flags.DEFINE_string("std2", "../data/COSMOS/std2.npy", "Standard deviation noise e2 (gal).")
-flags.DEFINE_float("sigma_gamma", 0.15, "Standard deviation of shear.")
 flags.DEFINE_float("spectral_norm", 1, "Amount of spectral normalization.")
-flags.DEFINE_string("variant", "EiffL", "Variant of model.")
-flags.DEFINE_string("model", "SmallUResNet", "Name of model.")
 flags.DEFINE_integer("map_size", 360, "Size of maps after cropping")
 flags.DEFINE_float("resolution", 0.29, "Resolution in arcmin/pixel")
 
 FLAGS = flags.FLAGS
 
 def forward_fn(x, is_training=False):
-  if FLAGS.model == 'SmallUResNet':
-    model = SmallUResNet(n_output_channels=1, variant=FLAGS.variant, deepmass=True)
-  elif FLAGS.model == 'MediumUResNet':
-    model = MediumUResNet()
-  else:
-    raise NotImplementedError
-  return model(x, jnp.zeros((1, 1, 1, 1)), is_training=is_training)
+  model = UResNet18(n_output_channels=1)
+  return model(x, condition=None, is_training=is_training)
 
 def lr_schedule(step):
   """Linear scaling rule optimized for 90 epochs."""
@@ -74,16 +62,16 @@ ks93inv = jax.vmap(ks93inv)
 
 def main(_):
   # Make the network
-  model = hk.transform_with_state(forward_fn)
+  model = hk.without_apply_rng(hk.transform_with_state(forward_fn))
 
   if FLAGS.spectral_norm > 0:
-    sn_fn = hk.transform_with_state(lambda x: CustomSNParamsTree(ignore_regex='[^?!.]*b$',
+    sn_fn = hk.transform_with_state(lambda x: SNParamsTree(ignore_regex='[^?!.]*b$|[^?!.]*offset$',
                                                               val=FLAGS.spectral_norm)(x))
 
   # Initialisation
-  optimizer = optix.chain(
-      optix.adam(learning_rate=FLAGS.learning_rate),
-      optix.scale_by_schedule(lr_schedule)
+  optimizer = optax.chain(
+      optax.adam(learning_rate=FLAGS.learning_rate),
+      optax.scale_by_schedule(lr_schedule)
   )
 
   rng_seq = hk.PRNGSequence(42)
@@ -91,7 +79,6 @@ def main(_):
   params, state = model.init(next(rng_seq),
                              jnp.zeros((1, FLAGS.map_size, FLAGS.map_size, 1)),
                              is_training=True)
-
   opt_state = optimizer.init(params)
 
   if FLAGS.spectral_norm > 0:
@@ -103,29 +90,28 @@ def main(_):
   std1 = jnp.expand_dims(jnp.load(FLAGS.std1).astype('float32'), 0)
   std2 = jnp.expand_dims(jnp.load(FLAGS.std2).astype('float32'), 0)
 
-  # Training loss
-  def loss_fn(params, state, rng_key, batch):
-    key1, key2, key3 = jax.random.split(rng_key, 3)
+  @jax.jit
+  def preprocess_batch(rng_key, batch):
+    """ Creates a noisy KS map as an input to the model
+    """
+    key1, key2 = jax.random.split(rng_key, 2)
     # Preprocess the batch for deep mass, i.e. apply KS, add noise, mask, and
     # do inverse Kaiser-Squires
     input_map = batch['x'][...,0]
     g1, g2 = ks93inv(input_map, jnp.zeros_like(input_map))
     
     # Add Gaussian noise and mask
-    # cte std per pixel
-    #g1 = mask * (g1 + FLAGS.sigma_gamma*jax.random.normal(key1, g1.shape))
-    #g2 = mask * (g2 + FLAGS.sigma_gamma*jax.random.normal(key2, g2.shape))
-
-    # std = std/sqrt(n_gal)
     g1 = mask * (g1 + std1*jax.random.normal(key1, g1.shape))
     g2 = mask * (g2 + std2*jax.random.normal(key2, g2.shape))
 
+    ks_map = jnp.stack(ks93(g1, g2), axis=-1)
+    return ks_map, input_map
 
-    ks_map, _ = ks93(g1, g2)
-    ks_map = jnp.expand_dims(ks_map,-1)
-    
+  # Training loss
+  def loss_fn(params, state, rng_key, batch):
+    ks_map, input_map = preprocess_batch(rng_key, batch)
     # Apply model
-    res, state = model.apply(params, state, key3, ks_map, is_training=True)
+    res, state = model.apply(params, state, ks_map, is_training=True)
     loss = jnp.mean((input_map - res[..., 0])**2)
     return loss, state
 
@@ -133,7 +119,7 @@ def main(_):
   def update(params, state, sn_state, rng_key, opt_state, batch):
     (loss, state), grads = jax.value_and_grad(loss_fn, has_aux=True)(params, state, rng_key, batch)
     updates, new_opt_state = optimizer.update(grads, opt_state)
-    new_params = optix.apply_updates(params, updates)
+    new_params = optax.apply_updates(params, updates)
     if FLAGS.spectral_norm > 0:
       new_params, new_sn_state = sn_fn.apply(None, sn_state, None, new_params)
     else:
@@ -159,15 +145,14 @@ def main(_):
     if step%100==0:
         print(step, loss)
 
-    # if step%500==0:
-    #   # Running denoiser on a batch of images
-    #   batch, res, gs = score_fn(params, state, next(rng_seq), next(train), is_training=False)
-    #   summary_writer.image('score/target', onp.clip(batch['x'][0], 0, 0.1)*10., step)
-    #   summary_writer.image('score/input', onp.clip(batch['y'][0], 0, 0.1)*10., step)
-    #   summary_writer.image('score/score', res[0]+gs[0], step)
-    #   summary_writer.image('score/denoised', onp.clip(batch['y'][0] + batch['s'][0,:,:,0]**2 * (res[0]+gs[0]), 0, 0.1)*10., step)
-    #   summary_writer.image('score/gaussian_denoised', onp.clip(batch['y'][0] + batch['s'][0,:,:,0]**2 * gs[0], 0, 0.1)*10., step)
-    #   print(step)
+    if step%500==0:
+      # Running denoiser on a batch of images
+      ks_map, input_map = preprocess_batch(next(rng_seq), next(train))
+      res, _ = model.apply(params, state, ks_map, is_training=False)
+      summary_writer.image('deepmass/target', onp.clip(input_map[0], 0, 0.1)*10., step)
+      summary_writer.image('deepmass/input', onp.clip(ks_map[0], 0, 0.1)*10., step)
+      summary_writer.image('deepmass/output', onp.clip(res[0], 0, 0.1)*10.   , step)
+      print(step)
 
     if step%5000 ==0:
       with open(FLAGS.output_dir+'/model-%d.pckl'%step, 'wb') as file:
